+++ config/connection.js
@@ -20,7 +20,7 @@ const sequelize = process.env.JAWSDB_URL
 cloudinary.config({
   cloud_name: process.env.CLOUD_NAME,
   api_key: process.env.API_KEY,
-  api_secret: process.envAPI_SECRET,
+  api_secret: process.env.API_SECRET,
 });
 
 module.exports = sequelize;
