+++ controllers/api/videoRoutes.js
@@ -3,13 +3,6 @@ const cloudinary = require("cloudinary").v2; // Setting up Cloudinery
 const { v4: uuidv4 } = require("uuid");
 // const { OpenVidu } = require("openvidu-node-client"); // using OpenVidu for WebRTC
 
-// Initialize Cloudinary
-// Replace with our cloud name when we get it HERE
-cloudinary.config({
-  cloud_name: "your_cloud_name",
-  api_key: "your_api_key",
-  api_secret: "your_api_secret",
-});
 
 // Initialize OpenVidu, replace this with out open video config??
 // WEB RTC uses this config option i think
@@ -72,5 +65,7 @@ router.post("/upload-video", async (req, res) => {
     res.status(500).json({ error: "Failed to upload video" });
   }
 });
+
+
 
 module.exports = router;
