+++ package.json
@@ -20,8 +20,10 @@
   },
   "homepage": "https://github.com/Chapjae/sillyCode_valley#readme",
   "dependencies": {
+    "@cloudinary/url-gen": "^1.11.1",
     "bcrypt": "^5.0.0",
     "cloudinary": "^1.40.0",
+    "cloudinary-core": "^2.13.0",
     "connect-session-sequelize": "7.1.7",
     "dotenv": "16.3.1",
     "express": "^4.18.2",
