+++ views/homepage.hbs
@@ -6,3 +6,4 @@
     <button type="button" class="btn-dark">Create A Video</button>
   </div>
 </div>
+<script src="/homepage.js"></script>
