+++ views/layouts/main.handlebars
@@ -1,9 +1,10 @@
+<!DOCTYPE html>
 <html lang="en">
 
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <meta http-equiv="Content-Type" content="application/javascript"/>
+    <meta http-equiv="Content-Type" content="application/javascript" />
     <title>Silly Code Valley!</title>
     <link
       href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.1/dist/css/bootstrap.min.css"
@@ -17,20 +18,23 @@
       crossorigin="anonymous"
     ></script>
     <link rel="stylesheet" href="/css/style.css" />
-<script src="https://upload-widget.cloudinary.com/global/all.js" type="text/javascript"></script>
+    <script
+      src="https://upload-widget.cloudinary.com/global/all.js"
+      type="text/javascript"
+    ></script>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/cloudinary-core/2.13.0/cloudinary-core.min.js" integrity="sha512-UJA1iBgFi7BUtsMcnT53Q9Utew/JgJNvrN+/lcMc3G2BmIzrUo0zR7lxjWsvlUWKWbHTBybGdGPj4UdUdIbT2g==" crossorigin="anonymous" referrerpolicy="no-referrer"></script>
   </head>
 
   <body>
     <div class="flex-column justify-space-around">
       <header class="display-flex justify-space-between align-center p-2">
-        <h1>🍿
-          <span role="img" aria-label="improv">🎬 🍿 </span>
-          <h1>IMPROV MASTERS!</h1>
+        <h1>
+          <span role="img" aria-label="improv">🍿🎬 🍿 </span>
+          <p>IMPROV MASTERS!</p>
         </h1>
         <nav>
           <!-- Conditionally render login or logout links -->
           {{#if logged_in}}
-            <a href="/profile">profile</a>
             |
             <button class="no-button" id="logout">logout</button>
           {{else}}
@@ -42,17 +46,15 @@
         <!-- Render the sub layout -->
         {{{body}}}
       </main>
-      <footer>
-
-      </footer>
     </div>
 
     <!-- Render script for logged in users only -->
     {{#if logged_in}}
       <script src="/js/logout.js"></script>
     {{/if}}
-    
-    {{!-- <script src="/script.js"></script> --}}
+    <footer>
+      we made this, we are very cool and smrt and awesome!
+    </footer>
   </body>
 
 </html>
